test(android): promote shared fakes to :core:client test fixtures
