--- apps/android/app/src/test/kotlin/social/waddle/android/TestRecords.kt
@@ -1,149 +1,0 @@
-package social.waddle.android
-
-import social.waddle.android.client.auth.WaddleSessionInfo
-import social.waddle.client.ffi.WaddleArchivedMessage
-import social.waddle.client.ffi.WaddleMamPage
-import social.waddle.client.ffi.WaddleMessage
-
-/** Fixture builders for the wide FFI records: overrides via named args. */
-
-fun testSessionInfo(
-    sessionId: String = "sess-1",
-    username: String = "icepuma",
-    jid: String = "icepuma@waddle.test",
-    isExpired: Boolean = false,
-): WaddleSessionInfo = WaddleSessionInfo(
-    sessionId = sessionId,
-    username = username,
-    avatarUrl = null,
-    xmppLocalpart = username,
-    jid = jid,
-    xmppWebsocketUrl = "wss://waddle.test/xmpp",
-    linkPreviewMediaOrigin = null,
-    isExpired = isExpired,
-    expiresAt = null,
-)
-
-fun testMamPage(
-    messages: List<WaddleArchivedMessage> = emptyList(),
-    firstId: String? = messages.firstOrNull()?.mamId,
-    lastId: String? = messages.lastOrNull()?.mamId,
-    isComplete: Boolean = false,
-): WaddleMamPage = WaddleMamPage(
-    messages = messages,
-    firstId = firstId,
-    lastId = lastId,
-    isComplete = isComplete,
-)
-
-fun testMessage(
-    id: String? = "msg-1",
-    from: String? = "alice@waddle.test",
-    to: String? = "icepuma@waddle.test",
-    body: String? = "hello",
-    messageType: String = "chat",
-    timestamp: String? = null,
-    stanzaId: String? = null,
-    originId: String? = null,
-    isMuc: Boolean = false,
-    thread: String? = null,
-): WaddleMessage = WaddleMessage(
-    id = id,
-    from = from,
-    to = to,
-    body = body,
-    subject = null,
-    messageType = messageType,
-    timestamp = timestamp,
-    stanzaId = stanzaId,
-    stanzaIdBy = null,
-    stanzaIds = emptyList(),
-    originId = originId,
-    replacesId = null,
-    retractsId = null,
-    retractionId = null,
-    isRetracted = false,
-    moderationTargetId = null,
-    moderatedBy = null,
-    moderationReason = null,
-    reactionTargetId = null,
-    reactionEmojis = emptyList(),
-    chatState = null,
-    displayedMarkerRequested = false,
-    displayedMarkerId = null,
-    isMuc = isMuc,
-    thread = thread,
-    parentThreadId = null,
-    markupSpans = emptyList(),
-    broadcastMention = null,
-    mentionUris = emptyList(),
-    references = emptyList(),
-    forumPostKind = null,
-    forumTitle = null,
-    isSticker = false,
-    linkPreviews = emptyList(),
-    pinEvent = null,
-    callThreadEnded = null,
-    carbon = null,
-    replyToId = null,
-    replyToSender = null,
-    replyFallbackStart = null,
-    replyFallbackEnd = null,
-    callThread = null,
-    sharedFiles = emptyList(),
-    mdsDisplayed = null,
-)
-
-fun testArchivedMessage(
-    mamId: String = "mam-1",
-    id: String? = null,
-    stanzaId: String? = null,
-    originId: String? = null,
-    timestamp: String? = "2026-07-15T10:00:00Z",
-    from: String? = "general@muc.waddle.test/alice",
-    to: String? = "icepuma@waddle.test",
-    messageType: String = "groupchat",
-    body: String? = "hello",
-): WaddleArchivedMessage = WaddleArchivedMessage(
-    mamId = mamId,
-    queryId = null,
-    id = id,
-    stanzaId = stanzaId,
-    stanzaIdBy = null,
-    stanzaIds = emptyList(),
-    originId = originId,
-    timestamp = timestamp,
-    from = from,
-    to = to,
-    messageType = messageType,
-    body = body,
-    subject = null,
-    replacesId = null,
-    retractsId = null,
-    retractionId = null,
-    isRetracted = false,
-    moderationTargetId = null,
-    moderatedBy = null,
-    moderationReason = null,
-    reactionTargetId = null,
-    reactionEmojis = emptyList(),
-    thread = null,
-    parentThreadId = null,
-    replyToId = null,
-    replyToSender = null,
-    replyFallbackStart = null,
-    replyFallbackEnd = null,
-    markupSpans = emptyList(),
-    broadcastMention = null,
-    mentionUris = emptyList(),
-    references = emptyList(),
-    forumPostKind = null,
-    forumTitle = null,
-    isSticker = false,
-    authorRealJid = null,
-    callThread = null,
-    callThreadEnded = null,
-    sharedFiles = emptyList(),
-    linkPreviews = emptyList(),
-    callEvent = null,
-)
